distiller/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.4
django-on-heroku==1.1.2
django-rest-knox==4.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Keras-Preprocessing==1.1.2
numpy==1.19.4
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
